requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
seaborn>=0.13.0
matplotlib>=3.7.0
statsmodels>=0.14.0
openpyxl>=3.1.2
pillow>=10.0.0
